moved serializers to directory with file for each class
